Increased size of RecieveMessages buffer from 1024 to 32678. 1024 was only grabbing about 4 parts of the 5 or 6 parts in the GetVersion response message before it returned resulting in an incomplete message being forwarded to ProcessMessage function
Moved verified functions to a new Requests.cs file

Verified return types of the following requests:
GetVersion
GetStats
GetVideoSettings
GetStreamServiceSettings
GetSceneList
GetSceneTransitionList
GetSourceFilterList
GetSourceFilterDefaultSettings
GetSceneItemList
GetOutputList
GetStreamStatus
GetRecordStatus

+++ COBWEBS-Client/Responses/STRUCT_GET_STATS.cs
@@ -13,11 +13,11 @@ namespace COBWEBS_Client.Responses
 		public double availableDiskSpace { get; set; }
 		public double activeFps { get; set; }
 		public double averageFrameRenderTime { get; set; }
-		public double renderSkippedFrames { get; set; }
-		public double renderTotalFrames { get; set; }
-		public double outputSkippedFrames { get; set; }
-		public double outputTotalFrames { get; set; }
-		public double webSocketSessionIncomingMessages { get; set; }
-		public double webSocketSessionOutgoingMessages { get; set; }
+		public ulong renderSkippedFrames { get; set; }
+		public ulong renderTotalFrames { get; set; }
+		public ulong outputSkippedFrames { get; set; }
+		public ulong outputTotalFrames { get; set; }
+		public ulong webSocketSessionIncomingMessages { get; set; }
+		public ulong webSocketSessionOutgoingMessages { get; set; }
 	}
 }
