+++ COBWEBS-Client/Responses/STRUCT_GET_STREAM_STATUS.cs
@@ -11,10 +11,10 @@ namespace COBWEBS_Client.Responses
 		public bool outputActive { get; set; }
 		public bool outputReconnecting { get; set; }
 		public string outputTimecode { get; set; }
-		public double outputDuration { get; set; }
-		public double outputCongestion { get; set; }
-		public long outputBytes { get; set; }
-		public int outputSkippedFrames { get; set; }
-		public int outputTotalFrames { get; set; }
+		public ulong outputDuration { get; set; }
+		public ulong outputCongestion { get; set; }
+		public ulong outputBytes { get; set; }
+		public ulong outputSkippedFrames { get; set; }
+		public ulong outputTotalFrames { get; set; }
 	}
 }
